feat: Create triggers with cozy-harvest-lib ✨
