+++ package.json
@@ -38,7 +38,7 @@
     "cozy-client": "5.7.3",
     "cozy-doctypes": "1.32.2",
     "cozy-flags": "1.5.1",
-    "cozy-harvest-lib": "0.29.1",
+    "cozy-harvest-lib": "0.30.0",
     "cozy-scripts": "1.12.2",
     "date-fns": "1.30.1",
     "enzyme-adapter-react-15": "1.2.0",
