+++ src/components/KonnectorInstall.jsx
@@ -1,109 +1,35 @@
-import React from 'react'
-import { translate } from 'cozy-ui/react/I18n'
+import React, { PureComponent } from 'react'
+import { connect } from 'react-redux'
+
+import { TriggerManager } from 'cozy-harvest-lib'
+import flag from 'cozy-flags'
+
+import LegacyKonnectorInstall from './LegacyKonnectorInstall'
+import { getKonnector } from 'ducks/konnectors'
 import styles from '../styles/konnectorInstall'
 
-import AccountLoginForm from './AccountLoginForm'
-import DescriptionContent from './DescriptionContent'
-import KonnectorMaintenance from './KonnectorMaintenance'
-import KonnectorSuccess from './KonnectorSuccess'
-
-import { getKonnectorMessage, isKonnectorLoginError } from '../lib/konnectors'
-import ErrorDescription from './ErrorDescriptions'
-
-export const KonnectorInstall = props => {
-  const {
-    t,
-    account,
-    connector,
-    disableSuccessTimeout,
-    error,
-    fields,
-    handleConnectionSuccess,
-    queued,
-    isUnloading,
-    oAuthTerminated,
-    onCancel,
-    onDone,
-    onSubmit,
-    submitting,
-    success,
-    successMessage,
-    successMessages,
-    trigger,
-    allRequiredFieldsAreFilled,
-    displayAdvanced,
-    toggleAdvanced,
-    isFetching,
-    isValid,
-    dirty,
-    successButtonLabel,
-    maintenance,
-    lang
-  } = props
-  const hasLoginError = isKonnectorLoginError(error)
-  const hasErrorExceptLogin = !!error && !hasLoginError
-  const isRunningInQueue = queued && submitting
-
-  return (
-    <div className={styles['col-account-connection-content']}>
-      <div className={styles['col-account-connection-form']}>
-        {hasErrorExceptLogin && ErrorDescription({ t, error, connector })}
-        {(!error || hasLoginError) &&
-          !isRunningInQueue &&
-          !success &&
-          !maintenance && (
-            <DescriptionContent
-              title={t('account.config.title', { name: connector.name })}
-              messages={[getKonnectorMessage(t, connector, 'terms')]}
-              centerTitle
-            />
-          )}
-        {maintenance && maintenance.longTerm ? (
-          <KonnectorMaintenance
-            maintenance={maintenance}
-            lang={lang}
-            konnectorName={connector.name}
-          />
-        ) : !account || !success || hasLoginError ? (
-          <AccountLoginForm
-            connectorSlug={connector.slug}
-            konnectorName={connector.name}
-            disableSuccessTimeout={disableSuccessTimeout}
-            error={hasLoginError}
-            fields={fields}
-            isValid={isValid}
-            dirty={dirty}
-            isFetching={isFetching}
-            forceEnabled={!!error}
-            isOAuth={connector.oauth}
-            isUnloading={isUnloading}
-            oAuthTerminated={oAuthTerminated}
-            onSubmit={onSubmit}
-            submitting={submitting}
-            allRequiredFieldsAreFilled={allRequiredFieldsAreFilled}
-            displayAdvanced={displayAdvanced}
-            toggleAdvanced={toggleAdvanced}
-          />
-        ) : (
-          <KonnectorSuccess
-            connector={connector}
-            error={error}
+export class KonnectorInstall extends PureComponent {
+  render() {
+    const { account, konnector, onDone, onLoginSuccess } = this.props
+    return flag('harvest') ? (
+      <div className={styles['col-account-connection-content']}>
+        <div className={styles['col-account-connection-form']}>
+          <TriggerManager
             account={account}
-            folderId={trigger && trigger.message.folder_to_save}
-            handleConnectionSuccess={handleConnectionSuccess}
-            isRunningInQueue={isRunningInQueue}
-            isUnloading={isUnloading}
+            konnector={konnector}
+            onLoginSuccess={onLoginSuccess}
             onDone={onDone}
-            onCancel={onCancel}
-            success={success}
-            title={successMessage}
-            messages={successMessages}
-            successButtonLabel={successButtonLabel}
           />
-        )}
+        </div>
       </div>
-    </div>
-  )
+    ) : (
+      <LegacyKonnectorInstall {...this.props} />
+    )
+  }
 }
 
-export default translate()(KonnectorInstall)
+const mapStateToProps = (state, ownProps) => ({
+  konnector: getKonnector(state.cozy, ownProps.connector.slug)
+})
+
+export default connect(mapStateToProps)(KonnectorInstall)
